requirements.txt
asgiref==3.8.1
Django==5.1.4
django-extensions==3.2.3
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0